requirements.txt
requests==2.27.1
tencentcloud-sdk-python-lighthouse==3.0.550
tencentcloud-sdk-python-common==3.0.550
pytz==2019.3